Added timeline functionality
Players can now add questions to their timeline and the logic will check if it is a valid position.

Unlocked timleine and locked timeline also added and will be sorted each time a new question is added to the timeline.

Added lives where if someone answers wrong 3 times they are redirected back to index and game restarts